Start to add the Connected Weighted Graph
My plan is to use the Connected Weighted Graph to store the edges
between each node, with each weight representing the number of times
those nodes have been paired.

Perhaps somewhat confusingly, I'm not planning to store the CWG as a
graph, but rather as a tree. Each node in the tree will correspond to a
node in the graph along with a list of the weighted edges that connect
it to other nodes. Each edge will only be stored once: i.e., a
connection between A and B will be stored at the tree node for A or for
B, but not for both.

Very little of this is implemented yet. I am just sketching out the
interface and the initial data structures.